+++ src/view/launcherView.ts
@@ -1,15 +1,25 @@
 import * as vscode from 'vscode';
 
 import { InfoItem, ListItem, ScriptItem, ArgItem, AddArgItem } from './components';
-import { scriptService } from '../services';
+import { localScriptService, globalScriptService } from '../services';
 
-function getScripts(): ScriptItem[] | InfoItem[] {
-    const scriptItems = scriptService.getScript().map(v => new ScriptItem(v));
-    return scriptItems.length === 0 ? [new InfoItem(vscode.l10n.t('Drop a script to here'))] : scriptItems;
+const globalIcon = new vscode.ThemeIcon('globe');
+const localIcon = new vscode.ThemeIcon('folder');
+
+function getGroupdScripts(): ListItem[] {
+    let globalScripts = globalScriptService.getScript().map(v => new ScriptItem(v));
+    let grouped = [new ListItem(vscode.l10n.t('Global'), globalScripts, vscode.l10n.t('${length} scripts'), globalIcon, 'globalScriptList')];
+
+    if (localScriptService) {
+        let lcoalScripts = localScriptService.getScript().map(v => new ScriptItem(v));
+        grouped.push(new ListItem(vscode.l10n.t('Workspace'), lcoalScripts, vscode.l10n.t('${length} scripts'), localIcon, 'localScriptList'));
+    }
+
+    return grouped;
 }
 
 function getArgs(script: ScriptItem): (ArgItem | AddArgItem)[] {
-    const argItems: ArgItem | AddArgItem[] = script.script.args.map((v, i) => new ArgItem(script.script, i));
+    const argItems: (ArgItem | AddArgItem)[] = script.script.args.map((v, i) => new ArgItem(script.script, i));
     argItems.push(new AddArgItem(script.script));
     return argItems;
 }
@@ -26,17 +36,21 @@ class FileDragAndDropController implements vscode.TreeDragAndDropController<Scri
         // console.log(source)
     }
 
-    async handleDrop(target: ScriptItem | undefined, dataTransfer: vscode.DataTransfer, token: vscode.CancellationToken) {
+    async handleDrop(target: ListItem | ScriptItem | ArgItem | AddArgItem | undefined, dataTransfer: vscode.DataTransfer, token: vscode.CancellationToken) {
         // console.log('drop')
         // console.log(dataTransfer)
-        // console.log(target)
         const uriList = dataTransfer.get('text/uri-list');
         if (uriList) {
             const lines = (await uriList.asString()).split('\r\n');
             const pathList = lines
                 .filter(line => line.startsWith('file:///'))
-                .map(line => line.slice(7));
-            vscode.commands.executeCommand('slurm--.addScript', pathList);
+                .map(line => vscode.Uri.parse(line));
+            let isLocal = (target instanceof ListItem)
+                ? (target.title === vscode.l10n.t('Workspace'))
+                : (target instanceof ScriptItem || target instanceof ArgItem || target instanceof AddArgItem)
+                    ? (target.script.isLocal)
+                    : true;
+            vscode.commands.executeCommand('slurm--.addScript', pathList, isLocal);
         }
     }
 }
@@ -66,7 +80,7 @@ export class LauncherViewDataProvider implements vscode.TreeDataProvider<ScriptI
 
     getChildren(element?: ScriptItem | ListItem): Thenable<ListItem[] | ScriptItem[] | InfoItem[] | (ArgItem | AddArgItem)[]> {
         if (!element) {
-            return Promise.resolve(getScripts());
+            return Promise.resolve(getGroupdScripts());
         }
         if (element instanceof ListItem) {
             return Promise.resolve(element.children as ScriptItem[]);